backend/requirements.txt
flask
flask-jwt-extended
flask-cors
flask-sqlalchemy
psycopg2-binary
boto3
flask-bcrypt
python-dotenv